Create person Mixin (#745)
* Create person Mixin

* Move more functions to mixin

* Address PR comments

* Fix h2 tag

* Remove prettier ignore
